Including rules
I gave every chess piece a number indicator that shows to which side it belongs to. I had to to it this way, because I didn't differentiate between the color of the piece during the development of the interface. Every chess piece implements the same interface. With this number indicator in changed the ChessBoard.java file. Now every chess piece has its own set of rules for the movement. I also implemented the rule that no chess piece can take over another chess piece frome the same side