Changing Files
I changed the files ChessBoard.java and Position.java. Now every chess piece abides every rule that applies to it in the rule world. It is no more possible for any chess piece to jump over another chess piece. If they collide with another chess piece on their path, then the position of this collision will be their new destination. If there is a chess piece of the same side on that destination, then nothing will happen. If the chess piece is from the enemy then, the chess piece will overtake that position.