Source code
I extended the ChessBoard.java file with a round counter. Now it is only possible for the person whose turn it is to make a move. If the other player tries to make a move, nothing will happen. This was made possible by checking if the board constellation changes after making a move and accordingly changing the round counter.