Final source code
I added another label to the frame that now indicates whose turn it is right now. I also finally added a win condition. If either of the kings get overtaken, the game will end with a message that say which side has won. The game should be no different from a real game of chess right now. The game is finished.