Final version
I realized that I forgot about the promotion rule in a chess game. If either a black or a white pawn reaches the other end of the board, there would simply be no move left for them to make. That's why I now implemented the promotion rule. If the stated case occurres, then the pawn piece will be replaced by a queen piece.